+++ src/ocamlorg_frontend/ocamlorg_frontend.ml
@@ -7,9 +7,8 @@ let about () = About.render ()
 let academic_users users = Academic_users.render users
 let best_practices best_practices = Best_practices.render best_practices
 let books books = Books.render books
-let community ~workshops ~news = Community.render ~workshops ~news
+let community ~workshops ~meetups = Community.render ~workshops ~meetups
 let workshop ~videos workshop = Workshop.render ~videos workshop
-let events ~workshops ~meetups = Events.render ~workshops ~meetups
 let home () = Home.render ()
 let industrial_users users = Industrial_users.render users
 let learn ~papers ~release ~books = Learn.render ~papers ~release ~books
@@ -21,10 +20,8 @@ let blog ~featured ~rss ~rss_page ~rss_pages_number ~news =
 let news ~page ~pages_number news = News.render ~page ~pages_number news
 let news_post news = News_post.render news
 
-let opportunities ?search ?country opportunities =
-  Opportunities.render ?search ?country opportunities
-
-let opportunity opportunity = Opportunity.render opportunity
+let opportunities ?search ?location ~locations opportunities =
+  Opportunities.render ?search ?location ~locations opportunities
 
 let package_overview ~readme ~dependencies ~rev_dependencies ~homepages ~source
     package =
@@ -53,5 +50,8 @@ let releases ?search releases = Releases.render ?search releases
 let success_stories success_stories = Success_stories.render success_stories
 let success_story success_story = Success_story.render success_story
 let tutorial tutorial = Tutorial.render tutorial
-let carbon_footprint () = Carbon_footprint.render ()
+
+let page ~title ~description ~meta_title ~meta_description ~content =
+  Page.render ~title ~description ~meta_title ~meta_description ~content
+
 let not_found () = Not_found.render ()