+++ tool/ood-gen/lib/job.ml
@@ -1,25 +1,18 @@
 type metadata = {
   title : string;
   link : string;
-  description : string;
   location : string;
   company : string;
   company_logo : string;
-  country : string;
-  fullfilled : bool;
 }
 [@@deriving yaml]
 
 type t = {
-  id : int;
   title : string;
   link : string;
-  description_html : string;
   location : string;
   company : string;
   company_logo : string;
-  country : string;
-  fullfilled : bool;
 }
 
 let path = Fpath.v "data/jobs.yml"
@@ -29,27 +22,20 @@ let decode s =
   match yaml with
   | `O [ ("jobs", `A xs) ] ->
       Ok
-        (List.rev
-           (List.mapi
-              (fun i ->
-                Utils.decode_or_raise (fun x ->
-                    match metadata_of_yaml x with
-                    | Ok raw ->
-                        Ok
-                          {
-                            id = i;
-                            title = raw.title;
-                            link = raw.link;
-                            description_html =
-                              Omd.of_string raw.description |> Omd.to_html;
-                            location = raw.location;
-                            country = raw.country;
-                            company = raw.company;
-                            company_logo = raw.company_logo;
-                            fullfilled = raw.fullfilled;
-                          }
-                    | Error err -> Error err))
-              (List.rev xs)))
+        (List.map
+           (Utils.decode_or_raise (fun x ->
+                match metadata_of_yaml x with
+                | Ok raw ->
+                    Ok
+                      {
+                        title = raw.title;
+                        link = raw.link;
+                        location = raw.location;
+                        company = raw.company;
+                        company_logo = raw.company_logo;
+                      }
+                | Error err -> Error err))
+           xs)
   | _ -> Error (`Msg "expected a list of jobs")
 
 let parse = decode
@@ -61,19 +47,14 @@ let all () =
 let pp ppf v =
   Fmt.pf ppf
     {|
-  { id = %a
-  ; title = %a
+  { title = %a
   ; link = %a
-  ; description_html = %a
   ; location = %a
-  ; country = %a
   ; company = %a
   ; company_logo = %a
-  ; fullfilled = %a
   }|}
-    Fmt.int v.id Pp.string v.title Pp.string v.link Pp.string v.description_html
-    Pp.string v.location Pp.string v.country Pp.string v.company Pp.string
-    v.company_logo Fmt.bool v.fullfilled
+    Pp.string v.title Pp.string v.link Pp.string v.location Pp.string v.company
+    Pp.string v.company_logo
 
 let pp_list = Pp.list pp
 
@@ -81,15 +62,11 @@ let template () =
   Format.asprintf
     {|
 type t =
-  { id : int
-  ; title : string
+  { title : string
   ; link : string
-  ; description_html : string
   ; location : string
-  ; country : string
   ; company : string
   ; company_logo : string
-  ; fullfilled : bool
   }
   
 let all = %a